v4

+++ tools/test-tool/src/flash/StageSprite.h
@@ -4,7 +4,7 @@
 class StageSprite :public Sprite
 {
 public:
-    StageSprite::StageSprite(int initialCapacity) :Sprite(initialCapacity) {
+    StageSprite::StageSprite(short initialCapacity) :Sprite(initialCapacity) {
         ;
     }
 };
